[MCUX-75800]Fix ctimer capture example coverity check issue
 - Fixed CERT INT31-C issue.
 - Modify DEMO_PullGpioPin parameter level from int to uint8_t to
   avoid unsigned to signed conversion.

+++ _boards/frdmmcxa166/driver_examples/ctimer/capture/app.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 NXP
+ * Copyright 2024-2025 NXP
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -22,7 +22,7 @@
 void BOARD_InitHardware(void);
 void DEMO_InitCtimerInput(void);
 void DEMO_InitGpioPin(void);
-void DEMO_PullGpioPin(int level);
+void DEMO_PullGpioPin(uint8_t level);
 /*${prototype:end}*/
 
 #endif /* _APP_H_ */
